requirements.txt
nature-remo
tinydb
pandas
matplotlib
